requirements.txt
# Python ML System Dependencies
torch>=2.0.0
torchvision>=0.15.0
scikit-learn>=1.3.0
pandas>=2.0.0
numpy>=1.24.0
matplotlib>=3.7.0
seaborn>=0.12.0
jupyter>=1.0.0
mlflow>=2.5.0
joblib>=1.3.0
pydantic>=2.0.0
click>=8.0.0
rich>=13.0.0
tqdm>=4.65.0
xgboost>=1.7.0
fastapi>=0.100.0
uvicorn>=0.23.0
pillow>=10.0.0
pytest>=7.4.0
pytest-cov>=4.1.0
plotly>=5.15.0
requests>=2.31.0